+++ packages/qiskit/pyproject.toml
@@ -30,6 +30,7 @@ python = "^3.9.8"
 numpy = ">=1.22.0"
 quri-parts-circuit = "*"
 qiskit = "^0.39.4"
+pytest = "^7.2.0"
 
 [tool.poetry.group.dev.dependencies]
 quri-parts-circuit = {path = "../circuit", develop = true}
